backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
alembic
asyncpg
aiosqlite
redis
httpx
python-dotenv
google-cloud-storage
google-cloud-translate
google-cloud-vision
google-auth
google-generativeai
librosa
python-socketio[asyncio]
pydub
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.28.0
